+++ lib/rucaptcha/controller_helpers.rb
@@ -14,17 +14,6 @@ module RuCaptcha
 
     # Generate a new Captcha
     def generate_rucaptcha
-      return generate_rucaptcha1
-      code = RuCaptcha::Captcha.random_chars
-      session_val = {
-        code: code,
-        time: Time.now.to_i
-      }
-      RuCaptcha.cache.write(rucaptcha_sesion_key_key, session_val, expires_in: RuCaptcha.config.expires_in)
-      RuCaptcha::Captcha.create(code)
-    end
-
-    def generate_rucaptcha1
       res = RuCaptcha.create()
       session_val = {
         code: res[0],